requirements.txt
boto3
moto3
